face-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy>=2.0.0,<3.0.0
Pillow>=10.4.0,<13.0.0
# DeepFace can be heavy and is not yet broadly compatible on latest Python.
# The service gracefully falls back to deterministic embeddings when absent.
deepface==0.0.93; python_version < "3.13"
